Refactor: 테스트 코드 분리
다른 레포지토리 테스트를 위해 상속구조로 변경하였습니다.
