+++ src/test/java/org/prgrms/kdt/common/BaseRepositoryTest.java
@@ -1,0 +1,56 @@
+package org.prgrms.kdt.common;
+
+import static com.wix.mysql.EmbeddedMysql.anEmbeddedMysql;
+import static com.wix.mysql.ScriptResolver.classPathScript;
+import static com.wix.mysql.config.Charset.UTF8;
+import static com.wix.mysql.config.MysqldConfig.aMysqldConfig;
+import static com.wix.mysql.distribution.Version.v8_0_11;
+
+import com.wix.mysql.EmbeddedMysql;
+import javax.sql.DataSource;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.MethodOrderer;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
+import org.junit.jupiter.api.TestMethodOrder;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
+
+/**
+ * Created by yhh1056
+ * Date: 2021/08/30 Time: 8:17 오후
+ */
+@Disabled
+@SpringJUnitConfig
+@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
+@TestInstance(Lifecycle.PER_CLASS)
+@ContextConfiguration(classes = JdbcDataConfig.class)
+public class BaseRepositoryTest {
+
+    @Autowired
+    protected DataSource dataSource;
+
+    protected EmbeddedMysql embeddedMysql;
+
+    @BeforeAll
+    void setUp() {
+        var mysqldConfig = aMysqldConfig(v8_0_11)
+                .withCharset(UTF8)
+                .withPort(2215)
+                .withUser("test", "test1234!")
+                .withTimeZone("Asia/Seoul")
+                .build();
+        embeddedMysql = anEmbeddedMysql(mysqldConfig)
+                .addSchema("test-order_mgmt", classPathScript("schema.sql"))
+                .start();
+    }
+
+    @AfterAll
+    void cleanup() {
+        embeddedMysql.stop();
+    }
+
+}
