+++ src/test/java/org/prgrms/kdt/customer/CustomerJdbcRepositoryTest.java
@@ -1,102 +1,34 @@
 package org.prgrms.kdt.customer;
 
-import static com.wix.mysql.EmbeddedMysql.anEmbeddedMysql;
-import static com.wix.mysql.ScriptResolver.classPathScript;
-import static com.wix.mysql.config.Charset.UTF8;
-import static com.wix.mysql.config.MysqldConfig.aMysqldConfig;
-import static com.wix.mysql.distribution.Version.v8_0_11;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.everyItem;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.samePropertyValuesAs;
 import static org.hamcrest.core.Is.is;
 
-import com.wix.mysql.EmbeddedMysql;
-import com.zaxxer.hikari.HikariDataSource;
 import java.time.LocalDateTime;
 import java.util.UUID;
-import javax.sql.DataSource;
-import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.TestInstance.Lifecycle;
-import org.junit.jupiter.api.TestMethodOrder;
+import org.prgrms.kdt.common.BaseRepositoryTest;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
 /**
  * Created by yhh1056
  * Date: 2021/08/30 Time: 6:14 오후
  */
-@SpringJUnitConfig
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-@TestInstance(Lifecycle.PER_CLASS)
-class CustomerJdbcRepositoryTest {
-
-    @Configuration
-    @ComponentScan(
-            basePackages = {"org.prgrms.kdt.customer"}
-    )
-    static class Config {
-
-        @Bean
-        public DataSource dataSource() {
-            var dataSource = DataSourceBuilder.create()
-                    .url("jdbc:mysql://localhost:2215/test-order_mgmt")
-                    .username("test")
-                    .password("test1234!")
-                    .type(HikariDataSource.class)
-                    .build();
-
-//            dataSource.setMinimumIdle(100);
-//            dataSource.setMaximumPoolSize(1000);
-            return dataSource;
-        }
-
-        @Bean
-        public JdbcTemplate jdbcTemplate(DataSource dataSource) {
-            return new JdbcTemplate(dataSource);
-        }
-
-    }
+class CustomerJdbcRepositoryTest extends BaseRepositoryTest {
 
     @Autowired
     CustomerJdbcRepository customerJdbcRepository;
 
-    @Autowired
-    DataSource dataSource;
-
     Customer newCustomer;
 
-    EmbeddedMysql embeddedMysql;
-
     @BeforeAll
-    void setUp() {
-        newCustomer = new Customer(UUID.randomUUID(), "test-user", "test-user@email.com", LocalDateTime
-                .now());
-        var mysqldConfig = aMysqldConfig(v8_0_11)
-                .withCharset(UTF8)
-                .withPort(2215)
-                .withUser("test", "test1234!")
-                .withTimeZone("Asia/Seoul")
-                .build();
-        embeddedMysql = anEmbeddedMysql(mysqldConfig)
-                .addSchema("test-order_mgmt", classPathScript("schema.sql"))
-                .start();
-    }
-
-    @AfterAll
-    void cleanup() {
-        embeddedMysql.stop();
+    void beforeAll() {
+        newCustomer = new Customer(UUID.randomUUID(), "test-user", "test-user@email.com", LocalDateTime.now());
     }
 
     @Test
